Move mapStateToProps fn in components
