+++ src/pages/ShowMnemonic/ShowMnemonicOnboarding.tsx
@@ -1,14 +1,20 @@
 import { IonContent, IonPage, IonGrid, IonRow, IonButton, useIonViewWillEnter, IonCol } from '@ionic/react';
 import * as bip39 from 'bip39';
 import React, { useState } from 'react';
+import { connect } from 'react-redux';
 import { useLocation } from 'react-router';
 import type { RouteComponentProps } from 'react-router';
+import type { Dispatch } from 'redux';
 
 import Checkbox from '../../components/Checkbox';
 import Header from '../../components/Header';
 import PageDescription from '../../components/PageDescription';
 import WordList from '../../components/WordList';
+import { setIsBackupDone } from '../../redux/actions/appActions';
+import { addErrorToast } from '../../redux/actions/toastActions';
+import type { RootState } from '../../redux/types';
 import type { AppError } from '../../utils/errors';
+import { setSeedBackupFlag } from '../../utils/storage-helper';
 
 interface ShowMnemonicProps extends RouteComponentProps {
   // connected redux props
@@ -81,4 +87,20 @@ const ShowMnemonicOnboarding: React.FC<ShowMnemonicProps> = ({ backupDone, histo
   );
 };
 
-export default ShowMnemonicOnboarding;
+const mapStateToProps = (state: RootState) => {
+  return {
+    backupDone: state.app.backupDone,
+  };
+};
+
+const mapDispatchToProps = (dispatch: Dispatch) => {
+  return {
+    setIsBackupDone: (done: boolean) => {
+      setSeedBackupFlag(done);
+      dispatch(setIsBackupDone(done));
+    },
+    onError: (err: AppError) => dispatch(addErrorToast(err)),
+  };
+};
+
+export default connect(mapStateToProps, mapDispatchToProps)(ShowMnemonicOnboarding);
