+++ src/pages/TorProxy/index.tsx
@@ -1,11 +1,12 @@
 import { IonButton, IonCol, IonContent, IonGrid, IonInput, IonItem, IonPage, IonRow } from '@ionic/react';
 import React, { useState } from 'react';
-import { useDispatch } from 'react-redux';
+import { connect, useDispatch } from 'react-redux';
 import type { RouteComponentProps } from 'react-router';
 
 import Header from '../../components/Header';
 import { setTorProxy } from '../../redux/actions/settingsActions';
 import { addSuccessToast } from '../../redux/actions/toastActions';
+import type { RootState } from '../../redux/types';
 
 interface TorProxyProps extends RouteComponentProps {
   torProxy: string;
@@ -66,4 +67,10 @@ const TorProxy: React.FC<TorProxyProps> = ({ torProxy }) => {
   );
 };
 
-export default TorProxy;
+const mapStateToProps = (state: RootState) => {
+  return {
+    torProxy: state.settings.torProxy,
+  };
+};
+
+export default connect(mapStateToProps)(TorProxy);
