--- src/redux/containers/backupContainer.tsx
@@ -1,28 +1,0 @@
-// Main Flow Back Up Page Container
-import { connect } from 'react-redux';
-import type { Dispatch } from 'redux';
-
-import Backup from '../../pages/Backup';
-import type { AppError } from '../../utils/errors';
-import { setSeedBackupFlag } from '../../utils/storage-helper';
-import { setIsBackupDone } from '../actions/appActions';
-import { addErrorToast } from '../actions/toastActions';
-import type { RootState } from '../types';
-
-const mapStateToProps = (state: RootState) => {
-  return {
-    backupDone: state.app.backupDone,
-  };
-};
-
-const mapDispatchToProps = (dispatch: Dispatch) => {
-  return {
-    setIsBackupDone: (done: boolean) => {
-      setSeedBackupFlag(done);
-      dispatch(setIsBackupDone(done));
-    },
-    onError: (err: AppError) => dispatch(addErrorToast(err)),
-  };
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(Backup);
