--- src/redux/containers/liquidityProvidersContainer.tsx
@@ -1,13 +1,0 @@
-import { connect } from 'react-redux';
-
-import LiquidityProvider from '../../pages/LiquidityProvider';
-import type { RootState } from '../types';
-
-const mapStateToProps = (state: RootState) => {
-  return {
-    network: state.settings.network,
-    providers: state.tdex.providers,
-  };
-};
-
-export default connect(mapStateToProps)(LiquidityProvider);
