--- src/redux/containers/showMnemonicContainer.tsx
@@ -1,27 +1,0 @@
-import { connect } from 'react-redux';
-import type { Dispatch } from 'redux';
-
-import ShowMnemonic from '../../pages/ShowMnemonic';
-import type { AppError } from '../../utils/errors';
-import { setSeedBackupFlag } from '../../utils/storage-helper';
-import { setIsBackupDone } from '../actions/appActions';
-import { addErrorToast } from '../actions/toastActions';
-import type { RootState } from '../types';
-
-const mapStateToProps = (state: RootState) => {
-  return {
-    backupDone: state.app.backupDone,
-  };
-};
-
-const mapDispatchToProps = (dispatch: Dispatch) => {
-  return {
-    setIsBackupDone: (done: boolean) => {
-      setSeedBackupFlag(done);
-      dispatch(setIsBackupDone(done));
-    },
-    onError: (err: AppError) => dispatch(addErrorToast(err)),
-  };
-};
-
-export default connect(mapStateToProps, mapDispatchToProps)(ShowMnemonic);
